Display wallet options in modal (#300)
* implement dropdown menu

* defer popup layout

* invalidate window after page is changed

* change workflow key

 * fix collapsible in test page

* add new requirements to docker file

* rescale images. add and use change page function

* correctly scale icons on transactions page

+++ ui/modal_templates.go
@@ -627,7 +627,7 @@ func (m *ModalTemplate) passwordsMatch(editors ...*widget.Editor) bool {
 func (m *ModalTemplate) passwordStrength(th *decredmaterial.Theme, editors ...*widget.Editor) {
 	password := editors[0]
 	strength := (dcrlibwallet.ShannonEntropy(password.Text()) / 4.0)
-	m.passwordStgth.Progress = int(strength * 100)
+	m.passwordStgth.Progress = float32(strength * 100)
 	m.passwordStgth.Color = th.Color.Success
 }
